Added Minimum Supported Asterisk Version (Error). Fixed sccp_log warning in ast108.c
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@2764 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/pbx_impl/ast/ast108.c
@@ -1114,7 +1114,7 @@ static int sccp_wrapper_asterisk18_set_rtp_peer(PBX_CHANNEL_TYPE * ast, PBX_RTP_
 	sccp_device_t *d = NULL;
 	struct ast_sockaddr them;
 
-	sccp_log((DEBUGCAT_CODEC)) (VERBOSE_PREFIX_1 "SCCP: (sccp_channel_set_rtp_peer) format: %d\n", codecs);
+	sccp_log((DEBUGCAT_CODEC)) (VERBOSE_PREFIX_1 "SCCP: (sccp_channel_set_rtp_peer) format: %d\n", (int)codecs);
 	sccp_log((DEBUGCAT_CHANNEL | DEBUGCAT_RTP)) (VERBOSE_PREFIX_1 "SCCP: __FILE__\n");
 	if (!(c = CS_AST_CHANNEL_PVT(ast))) {
 		sccp_log((DEBUGCAT_RTP)) (VERBOSE_PREFIX_1 "SCCP: (sccp_channel_set_rtp_peer) NO PVT\n");
